+++ base/src/sign.rs
@@ -80,6 +80,8 @@ pub trait Signed {
     fn sign(&self) -> Sign;
 }
 
+// TODO(v0.4): add is_positive and is_negative to the trait, and replace all the places where this is used
+
 macro_rules! impl_abs_ops_prim {
     ($($signed:ty;)*) => {$( // this branch is only for float
         impl Abs for $signed {